backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0
pydantic[email]==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.20
greenlet==3.1.1
